+++ tools/fi.py
@@ -434,6 +434,7 @@ def exercise(example_code, expect_exception):
     class myvfsfile(apsw.VFSFile):
 
         def __init__(self, parent, filename, flags):
+            filename.parameters
             super().__init__(parent, filename, flags)
 
     vfsinstance = myvfs()